fix(rust, python): fix cse profile
